# docs/api/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "TrustChain Public API",
    "description": "External TrustChain developer API. Authenticate with an organization API key (`Authorization: Bearer tc_live_***` or `tc_test_***`).",
    "version": "0.1.0",
    "contact": {
      "name": "TrustChain Developer Platform"
    },
    "license": {
      "name": "Proprietary"
    }
  },
  "servers": [
    {
      "url": "https://api.trustchain.local/api/public/v1",
      "description": "Public API v1"
    }
  ],
...
  "security": [
    {
      "ApiKeyBearer": []
    }
  ],
  "paths": {
...
    "/certificates": {
      "post": {
        "tags": [
          "Certificates"
        ],
        "summary": "Issue certificate",
        "operationId": "createCertificate",
        "parameters": [
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "title",
                  "recipientName"
                ],
                "properties": {
                  "title": {
                    "type": "string"
                  },
                  "recipientName": {
                    "type": "string"
                  },
                  "description": {
                    "type": "string",
                    "nullable": true
                  },
                  "recipientEmail": {
                    "type": "string",
                    "format": "email",
                    "nullable": true
                  },
                  "documentId": {
                    "type": "string",
                    "format": "uuid",
                    "nullable": true
                  },
                  "templateId": {
                    "type": "string",
                    "format": "uuid",
                    "nullable": true
                  },
                  "expiresAt": {
                    "type": "string",
                    "format": "date-time",
                    "nullable": true
                  },
                  "metadata": {
                    "type": "object",
                    "additionalProperties": true
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "certificate": {
                      "type": "object"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/certificates/{id}": {
      "get": {
        "tags": [
          "Certificates"
        ],
        "summary": "Get certificate",
        "operationId": "getCertificate",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Certificate",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "certificate": {
                      "type": "object"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "ApiKeyBearer": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "API key",
        "description": "Organization API key (`tc_live_…` / `tc_test_…`) or service-account secret (`sa_sec_…`)."
      }
    },
...
  }
}